feat: Initial set up of the project
It has been initialized the project. The basic components that will display
the information of the books has been added. Bulma will be used as
CSS framework and a preprocessor has been added as well (Sass).
It has been added because the use a variables is really powerful and,
about Bulma, I wanted to give it a try since it is an interesting and
cool CSS framework.

The different files of the project has been divided depending on its
function inside the src folder. Components will contain the partials
of the project like the component that will display a book or the list
of books. Containers will include the main components of the project,
like the books overview or the search section. Images will contain the
different assets. Utils will contain the BooksAPI file and other useful files
like helpers.