feat: Add search filter
In order to add the search filter and to maintain the entire status of
the books, the moveTo logic has been moved to the App component
so the search component has access to the moveTo function as well.
Thus if someone wants to add a book from the search to a shelf, now
he/she can. It has been added the search view but it has to be tested.